Depart: Updated error handling

+++ backend/controllers/departmentController.js
@@ -14,6 +14,12 @@ const createAllDepartments = asyncHandler(async (req, res) => {
   const data = await Promise.all(
     departmentsList.map(async (departmentInfo) => {
       const { name, email } = departmentInfo;
+
+      if (!name || !email) {
+        res.status(400);
+        throw new Error("Each department must have both an email and a name");
+      }
+
       let department = await Department.findOne({ name });
       if (!department) {
         department = await Department.create({
@@ -29,8 +35,10 @@ const createAllDepartments = asyncHandler(async (req, res) => {
     res.status(201).json({
       success: true,
       message: `Uploaded ${data.length} departments to the database`,
-      data,
     });
+  } else {
+    res.status(500);
+    throw new Error("Failed to creat all departments");
   }
 });
 
@@ -50,6 +58,12 @@ const getDepartments = asyncHandler(async (req, res) => {
 
 const getDepartmentById = asyncHandler(async (req, res) => {
   const { id } = req.params;
+
+  if (!id) {
+    res.status(400);
+    throw new Error("Invalid id");
+  }
+
   const department = await Department.findById(id);
   if (department) {
     res.status(200).json({
@@ -65,6 +79,10 @@ const getDepartmentById = asyncHandler(async (req, res) => {
 
 const updateDepartment = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  if (!id) {
+    res.status(400);
+    throw new Error("Invalid id");
+  }
   const department = await Department.findById(id);
   const { name, email, staff } = req.body;
 
@@ -87,6 +105,10 @@ const updateDepartment = asyncHandler(async (req, res) => {
 
 const deleteDepartment = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  if (!id) {
+    res.status(400);
+    throw new Error("Invalid id");
+  }
   const department = await Department.findByIdAndDelete(id);
   if (department) {
     res.status(200).json({
@@ -101,6 +123,11 @@ const deleteDepartment = asyncHandler(async (req, res) => {
 
 const addDepartment = asyncHandler(async (req, res) => {
   const { name, email, staff } = req.body;
+
+  if (!name || !email) {
+    res.status(400);
+    throw new Error("Department must have both an email and a name");
+  }
 
   const departmentExists = await Department.findOne({ name });
 
@@ -118,7 +145,6 @@ const addDepartment = asyncHandler(async (req, res) => {
     res.status(201).json({
       success: true,
       message: `${name} Department created successfully`,
-      data: department,
     });
   } else {
     res.status(400);
